Fix bug preventing screenshots from being taken
If `spaces.graph` has a value like `{"metrics":[]}`, then `Space#max_columns`
will return `nil` because `[].max == nil`.

To reproduce, create a new space, add a metric, and delete the metric. Then
`Space#max_columns` returns `nil`.

The error on sentry began only on July 5. Since then, it's been occurring every
hour when the task to take screenshots is run. My guess is that the task hadn't
been running for a while because of some other issue and that the deployment got
it running again.

The error message is

    ArgumentError: comparison of NilClass with 5 failed

https://sentry.io/guesstimate/guesstimate-production/issues/600210977