Upgrade Ruby to fix Logentries segfault
On Ruby 2.5.1 on Heroku, the app crashed on start-up because of a
segmentation fault. Here's the relevant portion of the error:

    2019-04-04T05:19:35.151598+00:00 app[web.1]: /app/vendor/ruby-2.5.1/bin/../lib/libruby.so.2.5(0x1569fb) [0x7f69626ca9fb]
    2019-04-04T05:19:35.155391+00:00 app[web.1]: /app/vendor/ruby-2.5.1/bin/../lib/libruby.so.2.5(0x21/app/vendor/bundle/ruby/2.5.0/gems/le-2.7.6/lib/le/host/http.rb:300: [BUG] Segmentation fault at 0x00007f6962757b8d
    2019-04-04T05:19:35.155394+00:00 app[web.1]: ruby 2.5.1p57 (2018-03-29 revision 63029) [x86_64-linux]
    2019-04-04T05:19:35.155396+00:00 app[web.1]:
    2019-04-04T05:19:35.155398+00:00 app[web.1]: -- Control frame information -----------------------------------------------
    2019-04-04T05:19:35.155400+00:00 app[web.1]: c:0030 p:---- s:0174 e:000173 CFUNC  :push
    2019-04-04T05:19:35.155402+00:00 app[web.1]: c:0029 p:0048 s:0169 e:000168 METHOD /app/vendor/bundle/ruby/2.5.0/gems/le-2.7.6/lib/le/host/http.rb:300
    2019-04-04T05:19:35.155404+00:00 app[web.1]: c:0028 p:0005 s:0165 e:000164 BLOCK  /app/vendor/bundle/ruby/2.5.0/gems/le-2.7.6/lib/le/host/http.rb:104 [FINISH]

Removing the `le` (Logentries) gem fixed this problem (see #194), but so did
upgrading to Ruby 2.5.5. It seems better to upgrade Ruby than to get rid of the
Logentries gem, since we might want to use Logentries. Also, if there's a bug
in the Ruby interpreter, other things besides Logentries could trigger it, and
it would be good to prevent that from happening.

For testing, I deployed the master branch and saw that there was the
stack trace I pasted above. When I visited
https://guesstimate-master.herokuapp.com/, I got Heroku's "Application
error" page. Then I deployed this branch. There was another segfault,
but this was just because the previous version was bad. When I deployed
this branch a second time, there was no segfault. And
https://guesstimate-master.herokuapp.com/ 404'd instead of 500'ing.

I only tested that the app doesn't crash; I didn't test any actual API
endpoints.