Upgrade to Ruby 2.7.2
Ruby 2.5.5 is incompatible with the heroku-20 stack. It would be
possible to install an older-but-still supported stack (namely
heroku-18), but then we'd have to do this all over again in a couple of
years.

It might be possible to upgrade only to Ruby 2.6, but while we're
upgrading things we might as well get to the most recent version that's
feasible, to make future upgrades less frequent or painful.

I decided not to upgrade to Ruby 3.0 because that has a ton of breaking
changes and would require upgrading a lot of gems and perhaps changing
app code as well.